cleanup: reorganize VenueData to Events domain
- Move VenueData from Shared to Events domain
- Update venue actions and repositories to use App\Data\Events\VenueData
- Remove duplicate root-level VenueData.php file
- Update test file to use correct VenueData namespace

Completes venue domain organization for both Models and Data.
