+++ app/Repositories/VenueRepository.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Repositories;
 
-use App\Data\Shared\VenueData;
+use App\Data\Events\VenueData;
 use App\Models\Events\Venue;
 use App\Repositories\Contracts\VenueRepositoryInterface;
 use App\Repositories\Support\BaseRepository;
